move type guards from types.ts to index.ts (serves as utils folder)
